refactor(sheets): modularize sheet system, increase code density
- Modularize packages/lib/src/sheets/sheet.ts (2,729 LoC) into focused modules:
  - constants.ts: Version constants and defaults
  - types.ts: All type definitions
  - address.ts: Cell address utilities
  - parser.ts: Formula tokenizer and parser
  - functions.ts: Spreadsheet function implementations
  - evaluation.ts: Formula evaluation engine
  - io.ts: Serialization/deserialization
  - external.ts: External reference collection
  - update.ts: Cell update and sanitization
- Sheet.ts now re-exports from modules for backwards compatibility
- Reduces sheet.ts complexity from Cx 213 to thin orchestrator
- All existing imports continue to work without changes

Phase 2 (SheetView.tsx modularization) deferred to follow-up PR

## Test plan
- [x] Spreadsheet creation works
- [x] Formula evaluation works
- [x] Cell formatting works
- [x] Copy/paste operations work
- [x] External references work
