+++ packages/lib/src/sheets/constants.ts
@@ -1,0 +1,13 @@
+/**
+ * @module @pagespace/lib/sheets/constants
+ * @description Version constants and default values for sheets
+ */
+
+export const SHEET_VERSION = 1;
+export const SHEET_DEFAULT_ROWS = 20;
+export const SHEET_DEFAULT_COLUMNS = 10;
+
+export const SHEETDOC_MAGIC = '#%PAGESPACE_SHEETDOC';
+export const SHEETDOC_VERSION = 'v1';
+
+export const LOCAL_PAGE_KEY = '__LOCAL_PAGE__';
